feat: remove circular imports issue crating constant file
