+++ packages/asset-packs/src/enums.ts
@@ -1,20 +1,8 @@
 import type { YGAlign, YGJustify } from '@dcl/ecs';
 import { getLatestVersionName } from './versioning';
 
-export const BaseComponentNames = {
-  ACTION_TYPES: 'asset-packs::ActionTypes',
-  ACTIONS: 'asset-packs::Actions',
-  COUNTER: 'asset-packs::Counter',
-  TRIGGERS: 'asset-packs::Triggers',
-  STATES: 'asset-packs::States',
-  COUNTER_BAR: 'asset-packs::CounterBar',
-  ADMIN_TOOLS: 'asset-packs::AdminTools',
-  VIDEO_SCREEN: 'asset-packs::VideoScreen',
-  REWARDS: 'asset-packs::Rewards',
-  TEXT_ANNOUNCEMENTS: 'asset-packs::TextAnnouncements',
-  VIDEO_CONTROL_STATE: 'asset-packs::VideoControlState',
-  SCRIPT: 'asset-packs::Script',
-} as const;
+// Re-export for backward compatibility
+export { BaseComponentNames } from './constants';
 
 export const ComponentName = {
   ACTION_TYPES: 'asset-packs::ActionTypes',
